storage-web/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pyjwt==2.8.1
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Frontend
streamlit==1.28.1
requests==2.31.0
pillow==10.1.0

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
